+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Upgrade `react-components` to [23.0.0](https://github.com/cucumber/react-components/releases/tag/v23.0.0)
 
 ## [21.10.1] - 2025-04-19
 ### Changed